src/core/operators/vid_vec_rep_resnet_requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --find-links=https://download.pytorch.org/whl/torch_stable.html vid_vec_rep_resnet_requirements.in
#

# For arm64 architecture
--find-links https://download.pytorch.org/whl/cpu
torch==2.2.0; platform_machine=='aarch64'
    # via
    #   -r vid_vec_rep_resnet_requirements.in
    #   torchvision
torchvision==0.17.0; platform_machine=='aarch64'
    # via -r vid_vec_rep_resnet_requirements.in

# For amd64 architecture
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cpu; platform_machine=='x86_64'
    # via
    #   -r vid_vec_rep_resnet_requirements.in
    #   torchvision
torchvision==0.17.0+cpu; platform_machine=='x86_64'
    # via -r vid_vec_rep_resnet_requirements.in

certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
filelock==3.13.1
    # via torch
fsspec==2024.2.0
    # via torch
idna==3.6
    # via requests
jinja2==3.1.3
    # via
    #   memray
    #   torch
linkify-it-py==2.0.3
    # via markdown-it-py
markdown-it-py[linkify,plugins]==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
markupsafe==2.1.5
    # via jinja2
mdit-py-plugins==0.4.0
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
memray==1.11.0
    # via -r vid_vec_rep_resnet_requirements.in
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   -r vid_vec_rep_resnet_requirements.in
    #   opencv-python-headless
    #   scipy
    #   torchvision
opencv-python-headless==4.9.0.80
    # via -r vid_vec_rep_resnet_requirements.in
pillow==10.2.0
    # via
    #   -r vid_vec_rep_resnet_requirements.in
    #   torchvision
pygments==2.17.2
    # via rich
pyinstrument==4.6.2
    # via -r vid_vec_rep_resnet_requirements.in
requests==2.31.0
    # via torchvision
rich==13.7.0
    # via
    #   memray
    #   textual
scipy==1.11.4
    # via -r vid_vec_rep_resnet_requirements.in
sympy==1.12
    # via torch
textual==0.50.0
    # via memray
typing-extensions==4.9.0
    # via
    #   textual
    #   torch
uc-micro-py==1.0.2
    # via linkify-it-py
urllib3==2.2.0
    # via requests
